+++ microwave/src/magnetron/oscillator.rs
@@ -128,10 +128,9 @@ pub struct StringSim<C> {
 impl<C: Controller> StringSim<C> {
     pub fn create_stage(&self) -> Stage<C::Storage> {
         let mut destination = self.destination.clone();
-        let num_samples_in_buffer =
-            (self.buffer_size_secs * f64::from(DEFAULT_SAMPLE_RATE)).ceil() as usize;
+        let num_samples_in_buffer = (self.buffer_size_secs * DEFAULT_SAMPLE_RATE).ceil() as usize;
 
-        let low_pass = OnePoleLowPass::new(0.0, DEFAULT_SAMPLE_RATE as f32);
+        let low_pass = OnePoleLowPass::new(0.0, DEFAULT_SAMPLE_RATE);
         let mut comb_filter = CombFilter::new(num_samples_in_buffer, 0.0, low_pass);
 
         let mut samples_processed = 0;
@@ -141,25 +140,24 @@ impl<C: Controller> StringSim<C> {
         let mut feedback = self.feedback.clone();
         let mut pluck_location = self.pluck_location.clone();
         Box::new(move |buffers, control| {
-            let frequency = frequency.next(control) as f32;
-            let cutoff = cutoff.next(control) as f32;
-            let feedback = feedback.next(control) as f32;
-            let pluck_location = pluck_location.next(control) as f32;
+            let frequency = frequency.next(control);
+            let cutoff = cutoff.next(control);
+            let feedback = feedback.next(control);
+            let pluck_location = pluck_location.next(control);
 
             comb_filter.set_feedback(-feedback);
             let low_pass = comb_filter.feedback_fn();
-            low_pass.set_cutoff(cutoff, DEFAULT_SAMPLE_RATE as f32);
+            low_pass.set_cutoff(cutoff, DEFAULT_SAMPLE_RATE);
             let intrinsic_delay = low_pass.intrinsic_delay_samples();
 
-            let num_samples_to_skip_back =
-                DEFAULT_SAMPLE_RATE as f32 / 2.0 / frequency - intrinsic_delay;
+            let num_samples_to_skip_back = DEFAULT_SAMPLE_RATE / 2.0 / frequency - intrinsic_delay;
 
             // Subtract 1.0 since the first skip-back sample is implicit
-            let offset = (num_samples_to_skip_back - 1.0) / num_samples_in_buffer as f32;
+            let offset = (num_samples_to_skip_back - 1.0) / num_samples_in_buffer as f64;
 
             let counter_wave_at =
-                (pluck_location.max(0.0).min(1.0) * DEFAULT_SAMPLE_RATE as f32 / 2.0 / frequency)
-                    .round() as usize;
+                (pluck_location.max(0.0).min(1.0) * DEFAULT_SAMPLE_RATE / 2.0 / frequency).round()
+                    as usize;
 
             buffers.write_1_read_0(&mut destination, control, || {
                 let input = if samples_processed > counter_wave_at {
@@ -172,7 +170,7 @@ impl<C: Controller> StringSim<C> {
                     samples_processed += 1;
                     0.0
                 };
-                f64::from(comb_filter.process_sample_fract(offset.max(0.0).min(1.0), input))
+                comb_filter.process_sample_fract(offset.max(0.0).min(1.0), input)
             })
         })
     }
